Use WithHooksAPIServiceTrait
