+++ layers/Engine/packages/component-model/src/Services/BasicServiceTrait.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PoP\ComponentModel\Services;
 
-use PoP\Hooks\HooksAPIInterface;
+use PoP\Hooks\Services\WithHooksAPIServiceTrait;
 use PoP\Root\Services\WithInstanceManagerServiceTrait;
 use PoP\Translation\TranslationAPIInterface;
 use Symfony\Contracts\Service\Attribute\Required;
@@ -12,19 +12,10 @@ use Symfony\Contracts\Service\Attribute\Required;
 trait BasicServiceTrait
 {
     use WithInstanceManagerServiceTrait;
+    use WithHooksAPIServiceTrait;
 
-    protected HooksAPIInterface $hooksAPI;
     protected TranslationAPIInterface $translationAPI;
 
-    #[Required]
-    final public function setHooksAPI(HooksAPIInterface $hooksAPI): void
-    {
-        $this->hooksAPI = $hooksAPI;
-    }
-    final protected function getHooksAPI(): HooksAPIInterface
-    {
-        return $this->hooksAPI;
-    }
     #[Required]
     final public function setTranslationAPI(TranslationAPIInterface $translationAPI): void
     {
