+++ layers/Engine/packages/loosecontracts/src/AbstractLooseContractResolutionSet.php
@@ -5,25 +5,18 @@ declare(strict_types=1);
 namespace PoP\LooseContracts;
 
 use PoP\Hooks\HooksAPIInterface;
+use PoP\Hooks\Services\WithHooksAPIServiceTrait;
 use PoP\Root\Services\AbstractAutomaticallyInstantiatedService;
 use PoP\Root\Services\WithInstanceManagerServiceTrait;
 
 abstract class AbstractLooseContractResolutionSet extends AbstractAutomaticallyInstantiatedService
 {
     use WithInstanceManagerServiceTrait;
+    use WithHooksAPIServiceTrait;
 
-    private ?HooksAPIInterface $hooksAPI = null;
     private ?LooseContractManagerInterface $looseContractManager = null;
     private ?NameResolverInterface $nameResolver = null;
 
-    final public function setHooksAPI(HooksAPIInterface $hooksAPI): void
-    {
-        $this->hooksAPI = $hooksAPI;
-    }
-    final protected function getHooksAPI(): HooksAPIInterface
-    {
-        return $this->hooksAPI ??= $this->instanceManager->getInstance(HooksAPIInterface::class);
-    }
     final public function setLooseContractManager(LooseContractManagerInterface $looseContractManager): void
     {
         $this->looseContractManager = $looseContractManager;
