+++ layers/Engine/packages/modulerouting/src/AbstractRouteModuleProcessor.php
@@ -5,22 +5,13 @@ declare(strict_types=1);
 namespace PoP\ModuleRouting;
 
 use PoP\Hooks\HooksAPIInterface;
+use PoP\Hooks\Services\WithHooksAPIServiceTrait;
 use PoP\Root\Services\WithInstanceManagerServiceTrait;
 
 abstract class AbstractRouteModuleProcessor
 {
     use WithInstanceManagerServiceTrait;
-
-    private ?HooksAPIInterface $hooksAPI = null;
-
-    final public function setHooksAPI(HooksAPIInterface $hooksAPI): void
-    {
-        $this->hooksAPI = $hooksAPI;
-    }
-    final protected function getHooksAPI(): HooksAPIInterface
-    {
-        return $this->hooksAPI ??= $this->instanceManager->getInstance(HooksAPIInterface::class);
-    }
+    use WithHooksAPIServiceTrait;
 
     /**
      * @return string[]
