+++ layers/Engine/packages/routing/src/AbstractRoutingManager.php
@@ -5,27 +5,18 @@ declare(strict_types=1);
 namespace PoP\Routing;
 
 use PoP\Hooks\HooksAPIInterface;
+use PoP\Hooks\Services\WithHooksAPIServiceTrait;
 use PoP\Root\Services\WithInstanceManagerServiceTrait;
 
 abstract class AbstractRoutingManager implements RoutingManagerInterface
 {
     use WithInstanceManagerServiceTrait;
+    use WithHooksAPIServiceTrait;
 
     /**
      * @var string[]|null
      */
     private ?array $routes = null;
-
-    private ?HooksAPIInterface $hooksAPI = null;
-
-    final public function setHooksAPI(HooksAPIInterface $hooksAPI): void
-    {
-        $this->hooksAPI = $hooksAPI;
-    }
-    final protected function getHooksAPI(): HooksAPIInterface
-    {
-        return $this->hooksAPI ??= $this->instanceManager->getInstance(HooksAPIInterface::class);
-    }
 
     /**
      * @return string[]
